# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Halaqaty API
  description: |
    REST API for the Halaqaty Quran memorization circle management platform.

    **Authentication:** Flutter uses Firebase Auth for sign-up, sign-in, password
    management, and ID-token refresh. API requests require a Firebase ID token in
    `Authorization: Bearer <token>` and, after backend session creation, the opaque
    current-device session identifier in `X-Halaqaty-Session-ID`. The backend never
    accepts passwords or issues Firebase ID or refresh tokens.

    **Base URL:** `https://api.halaqaty.app/api/v1`

    **WebSocket:** Real-time events (queue, session, chat) are delivered over WebSocket. See [ws_events.md](ws_events.md) for the full event catalogue.
  version: 1.0.0
  contact:
    name: Halaqaty
    url: https://halaqaty.app
servers:
  - url: https://api.halaqaty.app/api/v1
    description: Production
  - url: https://staging.api.halaqaty.app/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
    SessionId: []
...
paths:
...
  /circles/{circleId}/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/CircleId'
      - name: messageId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      operationId: deleteCircleMessage
      description: Delete a message (own messages within 10 min; teacher can delete any)
      tags: [Chat]
      summary: Delete a message (own messages within 10 min; teacher can delete any)
      parameters:
        - $ref: '#/components/parameters/RequiredIdempotencyKey'
      responses:
        '204':
          description: Message deleted
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase JWT
    SessionId:
      type: apiKey
      in: header
      name: X-Halaqaty-Session-ID
      description: Opaque backend-issued identifier for the current device session.
  parameters:
    CircleId:
      name: circleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RequiredIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Opaque retry key required for F-004 chat mutations; replaying different input returns 409.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, expired, or revoked Firebase ID token or backend session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks required role for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with the current circle state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Request rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: ERR_FORBIDDEN
            message:
              type: string
              description: Human-readable description
...